refactor: components 파일 타입스크립트로 변환
