+++ my-app/src/components/Episodes.tsx
@@ -1,7 +1,8 @@
-import React, { useState } from 'react';
+import { useState } from 'react';
 import styled from 'styled-components';
-import Rank from 'components/Rank';
-import PropTypes from 'prop-types';
+import Rank from '@/components/Rank';
+import { Episode } from '@/types';
+import noPosterImage from '@/assets/no_poster.png';
 
 const Title = styled.h2`
   text-transform: uppercase;
@@ -27,14 +28,14 @@ const PosterList = styled.ul`
   height: 100%;
   position: relative;
 `;
-const Item = styled.li`
+const Item = styled.li<{ isClicked: boolean }>`
   margin: 1rem auto;
   padding: 1rem;
   border-radius: 10px;
   width: 95%;
   cursor: pointer;
   transition: 0.5s background-color ease-in-out;
-  background-color: ${props => (props.isClicked ? '#70a1ff' : '#eee')};
+  background-color: ${(props) => (props.isClicked ? '#70a1ff' : '#eee')};
 `;
 
 const EpisodeTitle = styled.div`
@@ -63,8 +64,8 @@ const Overview = styled.div`
   margin: 0.5rem 0;
   color: black;
 `;
-const StillShot = styled.div`
-  background-image: url(${props => props.stillShotPoster});
+const StillShot = styled.div<{ stillShotPoster: string; isClicked: boolean; isExisted: boolean }>`
+  background-image: url(${(props) => props.stillShotPoster});
   background-repeat: no-repeat;
   width: 100%;
   height: 100%;
@@ -73,11 +74,11 @@ const StillShot = styled.div`
   left: 0;
   border-radius: 10px;
   transition: all 0.5s ease-in-out;
-  opacity: ${props => (props.isClicked ? 1 : 0)};
-  z-index: ${props => (props.isClicked ? 1 : 0)};
-  background-size: ${props => (props.isExisted ? '100%' : '90% 90%')};
+  opacity: ${(props) => (props.isClicked ? 1 : 0)};
+  z-index: ${(props) => (props.isClicked ? 1 : 0)};
+  background-size: ${(props) => (props.isExisted ? '100%' : '90% 90%')};
 `;
-const Order = styled.span`
+const Order = styled.span<{ current: boolean; isExisted: boolean }>`
   position: absolute;
   top: 5%;
   left: 5%;
@@ -85,16 +86,20 @@ const Order = styled.span`
   font-weight: 600;
   letter-spacing: 1px;
   transition: opacity 0.4s ease-in-out;
-  opacity: ${props => (props.current ? 1 : 0)};
-  color: ${props => (props.isExisted ? '#eee' : '#2d3436')};
+  opacity: ${(props) => (props.current ? 1 : 0)};
+  color: ${(props) => (props.isExisted ? '#eee' : '#2d3436')};
 `;
 
-function Episodes({ episodes }) {
+interface Props {
+  episodes: Episode[];
+}
+
+function Episodes({ episodes }: Props) {
   const [currentIndex, setCurrentIndex] = useState(0);
 
-  const handleClick = e => {
+  const handleClick = (e: React.MouseEvent<HTMLLIElement>) => {
     const { currentTarget } = e;
-    const _index = currentTarget.dataset.index;
+    const _index = Number(currentTarget.dataset?.index);
     setCurrentIndex(_index);
   };
 
@@ -110,10 +115,8 @@ function Episodes({ episodes }) {
                 key={episode.id}
                 data-index={index}
                 onClick={handleClick}
-                id={index}
+                id={String(index)}
                 isClicked={Number(index) === Number(currentIndex)}
-                currentIndex={currentIndex}
-                index={index}
               >
                 <EpisodeTitle>
                   <EpisodeNumber>{`Episode${episode.episode_number}`}</EpisodeNumber>{' '}
@@ -131,13 +134,11 @@ function Episodes({ episodes }) {
                 key={index}
                 isClicked={Number(index) === Number(currentIndex)}
                 stillShotPoster={
-                  episode.still_path
-                    ? `https://image.tmdb.org/t/p/w500${episode.still_path}`
-                    : require('../assets/no_poster.png')
+                  episode.still_path ? `https://image.tmdb.org/t/p/w500${episode.still_path}` : noPosterImage
                 }
-                isExisted={episode.still_path && true}
+                isExisted={!!episode.still_path && true}
               >
-                <Order current={Number(index) === Number(currentIndex)} isExisted={episode.still_path && true}>
+                <Order current={Number(index) === Number(currentIndex)} isExisted={!!episode.still_path && true}>
                   {episode.still_path
                     ? `S${episode.season_number} E${episode.episode_number}`
                     : 'Still shot Not Updated'}
@@ -150,21 +151,5 @@ function Episodes({ episodes }) {
     )
   );
 }
-
-Episodes.propTypes = {
-  episodes: PropTypes.arrayOf(
-    PropTypes.shape({
-      id: PropTypes.number.isRequired,
-      episode_number: PropTypes.number.isRequired,
-      name: PropTypes.string.isRequired,
-      air_date: PropTypes.string,
-      vote_average: PropTypes.number.isRequired,
-      vote_count: PropTypes.number.isRequired,
-      overview: PropTypes.string.isRequired,
-      still_path: PropTypes.string,
-      season_number: PropTypes.number.isRequired,
-    })
-  ).isRequired,
-};
 
 export default Episodes;
