+++ my-app/src/components/GlobalStyles.tsx
